Rename 'user_controller' to 'userController'
